feat: remove local API
Since the REST API has been running separately in production for years,
remove the last vestiges of having API endpoints in the web app itself,
removing a bunch of stuff no longer needed.

In case someone external is still hitting this undocumented API, 301 it
to where it should go.

The end goal should still be to move to a completely static web frontend
but this helps remove cruft from the web server version for now.